requirements.txt
#dracalvcp
#mmwranging
numpy
pyqt5
scipy
setuptools
twopilabs-sense-x1000
pyqtgraph
